Move balrog tests into balrog.rs
